Adds Retevis RT-3 code plug.
The FinDMR zone has every finnish DMR frequency with it
ordered by the frequency. TS1 has 244 and TS2 TG9. The
contact lists are made so that it should be easy to find
if there is traffic going on.

The Espoo-Kouvola zones each have one of the seven freqs.
Each programmed slot has only one talk group which it
listens. With these it should be possible to use any 70 cm
DMR repeater in Finland.

The analog zone has five simplex channels and 11 repeater
channels programmed into it. The scan list exists only for
the analog simplex channels.